+++ examples/release-explorer/cmd_main.go
@@ -1,0 +1,18 @@
+//go:build !glyph_demo_snap
+
+package main
+
+import (
+	"fmt"
+	"os"
+
+	tea "github.com/charmbracelet/bubbletea"
+)
+
+func main() {
+	p := tea.NewProgram(newModel(), tea.WithAltScreen())
+	if _, err := p.Run(); err != nil {
+		fmt.Fprintln(os.Stderr, "release-explorer:", err)
+		os.Exit(1)
+	}
+}